rename: 패키지 변경 - #1
- formatter 는 constant 보다는 util 에 가까우므로, util 패키지로 분리
